Documentation (#21)
* :sparkles: :pencil:

New documentation, although some parts are still WIP. Also added the
nightly version to the omitted build in the CI file.

* :bug:

The input files cannot be found by the builder in GitHub actions.

* :bug:

I removed an unncessary line that converted the classes to -1 and 1.
This was hard-coded and it's now left to the user.

* :bug:

I removed an unncessary line that converted the classes to -1 and 1.
This was hard-coded and it's now left to the user.

* :bug:

I recalled incorrectly, the problem is that the categories from the
CategorialArrays are indexed based, so the "class" 2 always exists.

* :pencil:

Documentation deployment to GitHub pages.

* :boom:

Remove support for CompatHelper. Compatibility can be enforced on
each clone of the repo.