Regression implementation (#22)
* :sparkles:

Added test and code for the new regressor type.

* :sparkles: :white_check_mark:

Initial training implementation for the regressors is working as expected
for a trivial benchmark.

* :sparkles:

Integration of the Least Square Support Vector Regressor with MLJ
has been successful. Examples and more documentation is now needed.

* :sparkles:

New documentation for the regressor implementation. Tests are already
implemented. A few more fixes are needed to the README.

* :pencil:

Some corrections to the documentation. Also fixed the wording in the
README.

* :pushpin:

Update to v0.4.