Refactoring of kernels (#25)
* :fire:

Started the refactoring to KernelFunctions.jl by removing some
unnecessary types and code.

* :bug:

Enforce the degree parameter to be an integer.

* :boom: :art:

Refactoring the computation of the kernel matrices for both training
and testing. Had to create some auxiliary functions for this purpose.

* :sparkles: :art:

The kernel refactorization for the classification formulation is complete.
We can now specify between three kernels, all supported by KernelFunctions.

* :art:

Split the utilities into a separate file for clarity and better
modularization.

* :art: :sparkles:

Included the utilities file to the complete package. Worked on the
docstrings for the utility functions, although these will not be seen
by the user, but rather by the developers.

* :sparkles: :art:

Successfully refactored the regression implementation. Partial tests
have passed.

* :sparkles: :white_check_mark:

All test pass. The problem was that kernelmatrix! expects vectors
instead of matrices, so another refactoring was needed. This time
kernelmatrix is used and no preallocation is needed now.

* :fire:

Removed unnecessary preallocations.

* :white_check_mark:

Added tests for the kernels available.

* :heavy_plus_sign: :heavy_minus_sign:

Added Documenter as a dependency. Removed BenchmarkTools and
DocumenterTools.

* :heavy_plus_sign: :pushpin:

Add Literate as a dependency. Fixed the current version to v0.6.

* :bug:

The documentation was sent to the main source code hierarchy.
The new path should fix that.

* :memo:

Export the SVM type so it gets added to the documentation.

* :green_heart:

Just use 64-bit architecture when using CI.

* :green_heart:

Revert to version 1 for documentation.

* :heavy_minus_sign:

Removed DocumenterTools dependency for documentation build step.

* :memo: :pushpin:

Working on a new example to showcase the usage of kernels.
Fixed the main dependencies, and the documentation dependencies
have been split from the main ones.

* :art:

Split the documentation creation into two steps. A Literate step
that gets done in the local machine and a make documentation step
that gets done on the CI servers.

* :memo:

Still working on the linear kernel example. The code runs perfectly,
just need to finish writing up the example itself.

* :memo:

Finished writing up the linear kernel example.

* :memo:

Examples are ready. This commit is large as I had to dabble with the
preprocessing of long-running examples.

* :memo: :green_heart:

Updated docstrings and added multithreading to the CI workflow.

* :memo:

Remove the returned nothing from code fences.