Multiclass classification implementation (#29)
* :white_check_mark:

Added the multiclass classification problem using the well-known
Iris dataset.

* :boom:

Correctly extracted the total number of classes. Still working on
a way to dispatch this problem.

* :sparkles:

Added a new method svmtrain_mc to handle the one-vs-one approach for
multiclass classification.

* :art:

By figuring out how the classes are encoded, I can now pass the labels
and implement the one-vs-one approach. The decoder stays out of the
if statement.

* :sparkles:

Working with the encoding of multiple classes.

* :recycle:

Using a function to find all the instances in an array that are
equal to some value.

* :sparkles:

The multiclass classification method is finished. By training several
binary classifiers we can then use these results for prediction.

* :sparkles:

Some refactoring to the training function for multiclass classification.
Added some comments, used views for the data matrix, and used a
collection to store the results of all the classifiers.

* :sparkles:

I found a way to use the information from the fit step when the prediction
step is called. The number of elements returned from the fitresult
tuple is enough.

* :sparkles:

The prediction step is almost done. By looping over the models and
their paramenters, I can decode the class pairs and build a matrix
of pooled predictions.

* :sparkles:

A new function which takes in the pooled predictions and uses a voting
scheme has been implemented.

* :sparkles:

The prediction step for the multiclass classification problem is
complete. A tie-break scheme is still needed.

* :memo: :white_check_mark:

Added some docstrings. Also added some clarifying comments on the
tests.

* :white_check_mark:

Recovering all tests to check the full test suite.

* :recycle:

No need to collect all the models, they are the same. This means that
the model has the same hyperparameters. Instead, just return a single
deep copy of the model.

* :memo:

More docstrings, comments, and documentation for clarity in the code.

* :recycle:

Using the in-place version of the broadcast function to save memory.
Removed the previous form of assignment.

* :pushpin:

New version to reflect the new multiclass implementation.

* :sparkles:

Working on a new example to showcase the multiclass classification
implementation.

* :recycle:

Some refactoring. I'm now passing a deep copy of the model each time
it is trained.

* :monocle_face:

Very bad performance on the multiclass classification implementation.

* :bug: :recycle:

Bad performance was fixed due to a bug inside the double loop.
I was looping over the same index. Also did some slight refactoring
just to improve code formatting.

* :art:

Using the auto-formatter for formatting code.

* :zap:

Perfect classification rate for the multiclass example.

* :sparkles:

The multiclass example is complete.

* :memo:

Building the final documentation integrating the multiclass example.

* :memo:

Final adjustments to the examples in the documentation.